requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.3
django-heroku==0.3.1
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
joblib==1.1.0
MarkupPy==1.14
nltk==3.7
numpy==1.22.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
tqdm==4.63.1
urllib3==1.26.9
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
